+++ src/frontend/apps/standalone_site/src/features/Contents/features/Video/components/Manage/VideoCreateForm.tsx
@@ -1,4 +1,5 @@
-import { Box, Text, TextArea, TextInput } from 'grommet';
+import { Field, Input } from '@openfun/cunningham-react';
+import { Box, Text, TextArea } from 'grommet';
 import { Alert } from 'grommet-icons';
 import { Nullable } from 'lib-common';
 import {
@@ -147,14 +148,21 @@ const VideoCreateForm = () => {
             display: 'block',
           }}
         >
-          <FormField
-            label={intl.formatMessage(messages.titleLabel)}
-            htmlFor="title-id"
-            name="title"
-            required
-          >
-            <TextInput size="1rem" name="title" id="title-id" />
-          </FormField>
+          <Field className="mb-s" fullWidth>
+            <Input
+              aria-label={intl.formatMessage(messages.titleLabel)}
+              fullWidth
+              label={intl.formatMessage(messages.titleLabel)}
+              name="title"
+              required
+              onChange={(e) => {
+                setVideo((value) => ({
+                  ...value,
+                  title: e.target.value,
+                }));
+              }}
+            />
+          </Field>
 
           {selectPlaylist}