+++ src/frontend/apps/standalone_site/src/features/Profile/features/AccountSettings/api/resetPassword.ts
@@ -1,11 +1,17 @@
 import { Maybe } from 'lib-common';
-import { fetchWrapper, useJwt } from 'lib-components';
+import { fetchResponseHandler, fetchWrapper, useJwt } from 'lib-components';
+
+export interface ResetPasswordError {
+  old_password?: string[];
+  new_password1?: string[];
+  new_password2?: string[];
+}
 
 export const resetPassword = async (
   currentPassword: Maybe<string>,
   newPassword: Maybe<string>,
   passwordValidation: Maybe<string>,
-) => {
+): Promise<void> => {
   const jwt = useJwt.getState().getJwt();
 
   const response = await fetchWrapper('/account/api/password/change/', {
@@ -21,7 +27,5 @@ export const resetPassword = async (
     method: 'POST',
   });
 
-  if (!response.ok) {
-    throw await response.json();
-  }
+  await fetchResponseHandler<void>(response);
 };